requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-firestore
google-cloud-aiplatform
ddtrace
python-dotenv
requests
pydantic
datadog